Refactor: Extract game logic into Game component
This commit refactors the Minesweeper game by extracting the core game logic into a new `Game` component. This improves code organization and reusability. The API route for ranking has also been removed and replaced with server actions.
